Solution to Calculate Average Temperature of a Week

+++ Calculate Average Temperature of a Week@@67877d61dc107486ad5c274d/code.c
@@ -2,7 +2,7 @@
 
 struct week{
     int days;
-    int tempp;
+    float tempp;
 };
 
 int main() {
